Better alternative to ContinuationPage-like classes
Make the `info` field optional on the normal classes and have a
`continuation` property that returns a Response using the
search.html.jinja template.
This fixes continuation pages missing the properties of the original
classes, causing the lack of channel shorts/related channels grouping
when the featured tab has a lot of root entries.